exp/upbox, exp/cmd/upbox: move out of exp directory
Now that upbox is used by the tests in cmd/upspin, it should no longer
be considered experimental.
